udn, unit tests: move newMultiHomedPod to the multihoming test utils
This function is commonly used by all secondary controllers (L2 and L3).
It should even be re-used in the future by the localnet controller, once
it is covered by unit tests.
